[update] logger, support straightly import logger to use
