+++ fastNLP/io/__init__.py
@@ -72,6 +72,8 @@ __all__ = [
 
     'ModelLoader',
     'ModelSaver',
+
+    'logger',
 ]
 
 from .embed_loader import EmbedLoader
@@ -81,3 +83,4 @@ from .model_io import ModelLoader, ModelSaver
 
 from .loader import *
 from .pipe import *
+from ._logger import *
